Add notifications

+++ bugbook/app/api/posts/[postId]/likes/route.ts
@@ -55,19 +55,43 @@ export async function POST(
     if (!loggedInUser)
       return Response.json({ error: "Unauthorised" }, { status: 401 });
 
-    await db.like.upsert({
-      where: {
-        userId_postId: {
+    const post = await db.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post)
+      return Response.json({ error: "Post not found" }, { status: 404 });
+
+    await db.$transaction([
+      db.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
+          },
+        },
+        create: {
           userId: loggedInUser.id,
           postId,
         },
-      },
-      create: {
-        userId: loggedInUser.id,
-        postId,
-      },
-      update: {},
-    });
+        update: {},
+      }),
+      ...(loggedInUser.id !== post.userId
+        ? [
+            db.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "like",
+              },
+            }),
+          ]
+        : []),
+    ]);
 
     return new Response();
   } catch (error) {
@@ -85,12 +109,32 @@ export async function DELETE(
     if (!loggedInUser)
       return Response.json({ error: "Unauthorised" }, { status: 401 });
 
-    await db.like.deleteMany({
-      where: {
-        userId: loggedInUser.id,
-        postId,
+    const post = await db.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       },
     });
+
+    if (!post)
+      return Response.json({ error: "Post not found" }, { status: 404 });
+
+    await db.$transaction([
+      db.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId,
+        },
+      }),
+      db.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "like",
+        },
+      }),
+    ]);
 
     return new Response();
   } catch (error) {
